api/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.2
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
